issue/3105-react-fixes isInteractive, graphic.jsx>image.jsx, empty partials
